generate embeddings
