+++ python/benchmarks/oxford_pet/train.py
@@ -17,14 +17,12 @@ import torchvision
 import torchvision.transforms as T
 from torchvision.transforms.functional import InterpolationMode
 from PIL import Image
-from torch import optim
 from torchdata.datapipes.iter import IterableWrapper
 from pytorch_lightning.loggers import TensorBoardLogger
-from torchvision.models.efficientnet import EfficientNet_B0_Weights
-from torchvision.models.resnet import ResNet50_Weights
 
 import lance
 import lance.pytorch.data
+from common import Classification, RawOxfordPetDataset, raw_collate_fn
 
 NUM_CLASSES = 38
 
@@ -59,125 +57,6 @@ class TrainTransform(torch.nn.Module):
         return self.transform(tensor)
 
 
-def raw_collate_fn(batch):
-    images = []
-    labels = []
-    for img, label in batch:
-        images.append(img)
-        labels.append(label)
-    return torch.stack(images), torch.tensor(labels)
-
-
-class RawOxfordPetDataset(torch.utils.data.Dataset):
-    """Build OxfordPet dataset from raw metadata.
-
-    The main difference between this and the one from torchvision is that
-    this one can directly read from cloud storage.
-    """
-
-    def __init__(
-        self, root: str, split: str = "trainval", transform: Optional[Callable] = None
-    ):
-        """Constructor.
-
-        Parameters
-        ----------
-        root : str
-            Root URI of the dataset, i.e., `s3://eto-public/datasets/oxford_pet/`.
-            We expect two sub-directories under the root: "annotations" and "images".
-        split : str
-            The split to load, either "trailval" or "test"
-        """
-        self.root = root
-        self.split = split
-        self.transform = transform
-
-        image_list_file = os.path.join(root, "annotations", f"{split}.txt")
-        fs, path = pyarrow.fs.FileSystem.from_uri(image_list_file)
-        self.fs = fs
-        self.images = []
-        self.labels = []
-        with fs.open_input_stream(path) as input:
-            content = input.readall().decode("utf-8")
-            for line in content.split("\n"):
-                comps = line.split()
-                if comps:
-                    self.images.append(comps[0])
-                    self.labels.append(torch.tensor(int(comps[1]), dtype=torch.long))
-
-    def __len__(self):
-        return len(self.images)
-
-    def __getitem__(self, idx):
-        image_uri = os.path.join(self.root, "images", f"{self.images[idx]}.jpg")
-        fs, path = pyarrow.fs.FileSystem.from_uri(image_uri)
-        with fs.open_input_stream(path) as fobj:
-            img = Image.open(io.BytesIO(fobj.readall())).convert("RGB")
-            if self.transform:
-                img = self.transform(img)
-            return img, self.labels[idx]
-
-
-def collate_fn(transform):
-    def _collate_fn(batch):
-        # TODO: Labels should be converted via torch.LanceDataset
-        labels = torch.from_numpy(batch[1])
-        # TODO: Image conversion should in torch.LanceDataset
-        images = [
-            transform(Image.open(io.BytesIO(data)).convert("RGB")) for data in batch[0]
-        ]
-        return torch.stack(images), labels
-
-    return _collate_fn
-
-
-class Classification(pl.LightningModule):
-    """Classification model to train"""
-
-    def __init__(
-        self,
-        model: torch.nn.Module = torchvision.models.efficientnet_b0(
-            num_classes=NUM_CLASSES
-        ),
-        learning_rate=0.1,
-        benchmark: Optional[str] = None,
-    ) -> None:
-        """Build a PyTorch classification model."""
-        super().__init__()
-        self.model = model
-        self.criterion = torch.nn.CrossEntropyLoss()
-        self.benchmark = benchmark
-        self.fit_start_time = 0
-        self.learning_rate = learning_rate
-
-    def on_fit_start(self) -> None:
-        super().on_fit_start()
-        self.fit_start_time = time.time()
-
-    def on_fit_end(self) -> None:
-        super().on_fit_end()
-        print(f"Training finished in {time.time() - self.fit_start_time} seconds")
-
-    def training_step(self, batch, batch_idx):
-        """
-        https://github.com/pytorch/vision/blob/main/references/classification/train.py
-        """
-        images, labels = batch
-        if self.benchmark == "io":
-            # only test I/O
-            pass
-        else:
-            output = self.model(images)
-            loss = self.criterion(output, labels)
-            self.log_dict({"loss": loss})
-            return loss
-
-    def configure_optimizers(self):
-        # Use hyperparameters from https://github.com/pytorch/vision/tree/main/references/classification
-        #
-        optimizer = optim.SGD(self.parameters(), lr=(self.learning_rate), momentum=0.9, weight_decay=1e-4)
-        return optimizer
-
 
 @click.command()
 @click.option("-b", "--batch_size", default=64, help="batch size", show_default=True)
@@ -187,13 +66,6 @@ class Classification(pl.LightningModule):
     "--num_workers",
     default=os.cpu_count(),
     help="set pytorch DataLoader number of workers",
-    show_default=True,
-)
-@click.option(
-    "-m",
-    "--model",
-    type=click.Choice(["resnet", "efficientnet"]),
-    default="resnet",
     show_default=True,
 )
 @click.option(
@@ -214,6 +86,7 @@ def train(
     num_workers: int,
     data_format,
 ):
+
     if model == "resnet":
         m = torchvision.models.resnet50(num_classes=NUM_CLASSES)
         transform = TrainTransform(crop_size=224)
@@ -250,9 +123,9 @@ def train(
     else:
         raise ValueError("Unsupported data format")
 
-    logger = TensorBoardLogger("tensorboard", name=f"oxford_pet_{model}")
+    logger = TensorBoardLogger("logs", name=f"oxford_pet_{model}")
 
-    model = Classification(benchmark=benchmark, model=m)
+    model = Classification(benchmark=benchmark, backbone=m)
     trainer = pl.Trainer(
         limit_train_batches=100,
         max_epochs=epochs,
@@ -263,6 +136,7 @@ def train(
     )
     trainer.tune(model, train_dataloaders=train_loader)
     trainer.fit(model=model, train_dataloaders=train_loader)
+    trainer.save_checkpoint("model.ckt")
 
 
 if __name__ == "__main__":
